Code Cleanup
Restructured folders
Separated files
Fixed tabs in help
