+++ Black Jack/include/main.h
@@ -1,0 +1,15 @@
+//
+//  main.cpp
+//  Black Jack
+//
+//  Created by Justin Oroz on 11/20/14.
+//  Copyright (c) 2014 Justin Oroz. All rights reserved.
+//
+
+#ifndef __Black_Jack__game__
+#define __Black_Jack__game__
+
+#include <stdio.h>
+#include "game.hpp"
+
+#endif /* defined(__Black_Jack__game__) */
